Explicitly use /usr/include/ncurses when building _curses and _cursespanel.
On oneiric and earlier, libreadline-dev would pull in libncurses5-dev which
provides /usr/include/curses.h; however, python3.1 only Build-Depends on
libncursesw5-dev (which is the proper library to use), but didn't actually have
any support for /usr/include/ncursesw/curses.h. I'm not sure it's a problem on
oneiric, but it means _curses fails on precise, so I'm just adding it for
oneiric as well.